Send dataLayer event when render is finished
* To check if this resolves/reduces GTM blocking issues.
* In order to "know" when all blocks are done, we need to render them in
a single React element using portals, instead of doing separate renders.
* Also emit event at start and finish of frontendIndex.js.
